LayoutTests/imported/w3c/web-platform-tests/navigation-api/navigation-methods/return-value/traverseTo-detach-cross-document-before-navigate-event-expected.txt



FAIL traverseTo() promise rejections when detaching an iframe before onnavigate (cross-document) assert_equals: expected 2 but got 1
